requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.19
pydantic==2.7.1
python-dotenv==1.0.1
pandas==2.2.2
python-multipart==0.0.9
